refactor: rename section-related enum and update imports
