+++ src/components/common/section-templates/dual-images-bottom/DualImagesBottom.test.tsx
@@ -1,6 +1,6 @@
 import { DualImagesBottom, DualImagesBottomProps } from './DualImagesBottom';
 import { createImagesBottomTestSuite } from '../shared/test-utils/imagesBottomTestFactory';
-import { SectionMode } from '@/types/common/program-sections';
+import { SectionMode } from '@/types/common/sections';
 
 jest.mock('../shared/title-description-section/TitleDescriptionSection', () => {
     const { mockTitleDescriptionSection } = require('../shared/test-utils/imagesBottomTestMocks');
