+++ src/components/common/section-templates/quad-images-bottom/QuadImagesBottom.test.tsx
@@ -1,6 +1,6 @@
 import { QuadImagesBottom, QuadImagesBottomProps } from './QuadImagesBottom';
 import { createImagesBottomTestSuite } from '../shared/test-utils/imagesBottomTestFactory';
-import { SectionMode } from '@/types/common/program-sections';
+import { SectionMode } from '@/types/common/sections';
 
 jest.mock('../shared/title-description-section/TitleDescriptionSection', () => {
     const { mockTitleDescriptionSection } = require('../shared/test-utils/imagesBottomTestMocks');
