+++ src/components/common/section-templates/shared/title-description-cards-section/TitleDescriptionCardsSection.tsx
@@ -1,6 +1,6 @@
 import cn from 'classnames';
 import { TitleDescriptionCard } from './TitleDescriptionCard';
-import { SectionMode, SectionTemplate } from '@/types/common/program-sections';
+import { SectionMode, SectionTemplate } from '@/types/common/sections';
 import { getProgramSectionTemplateMaxGroupCount } from '@/utils/functions/program-section-template-validation/programSectionTemplateValidation';
 import styles from './TitleDescriptionCardsSection.module.scss';
