+++ src/pages/admin/programs/components/programs-page-modals/add-section-modal/AddSectionModal.tsx
@@ -2,7 +2,7 @@ import { useState } from 'react';
 import { Modal } from '@/components/common/modal/Modal';
 import { Button } from '@/components/admin/button/Button';
 import { PROGRAMS_TEXT } from '@/const/admin/programs';
-import { SectionMode, SectionTemplate } from '@/types/common/program-sections';
+import { SectionMode, SectionTemplate } from '@/types/common/sections';
 import { Swiper } from '@/components/public/swiper/Swiper';
 import { renderProgramSection } from '@/utils/functions/render-program-section';
 import { MockQuestions } from '@/utils/mock-data/public/programs-page';
