+++ src/pages/admin/programs/components/programs-page-modals/program-modal/ProgramModal.tsx
@@ -8,7 +8,8 @@ import {
     SectionCancelActionType,
 } from '@/types/admin/programs';
 import { VisibilityStatus, PendingAction, ModalMode } from '@/types/admin/common';
-import { CreateHippotherapyProgramSectionDto, SectionTemplate } from '@/types/common/program-sections';
+import { CreateHippotherapyProgramSectionDto } from '@/types/common/program-sections';
+import { SectionTemplate } from '@/types/common/sections';
 import { PROGRAMS_TEXT } from '@/const/admin/programs';
 import { COMMON_TEXT_ADMIN } from '@/const/admin/common';
 import { ProgramsApi } from '@/services/api/admin/programs/programs-api';
